add "open" option to node api to allow opening in browser
