+++ lib/Server.js
@@ -27,6 +27,9 @@ const killable = require('killable');
 const del = require('del');
 const chokidar = require('chokidar');
 
+const open = require('opn');
+const createDomain = require('./utils/createDomain');
+
 const express = require('express');
 
 const compress = require('compression');
@@ -76,6 +79,22 @@ const STATS = {
   errors: true,
   errorDetails: false
 };
+
+function openBrowser(uri, options, log) {
+  let openOptions = {};
+  let openMessage = 'Unable to open browser';
+
+  if (typeof options.open === 'string') {
+    openOptions = { app: options.open };
+    openMessage += `: ${options.open}`;
+  }
+
+  open(uri + (options.openPage || ''), openOptions).catch(() => {
+    log.warn(
+      `${openMessage}. If you are running in a headless environment, please do not use the --open flag`
+    );
+  });
+}
 
 function Server (compiler, options = {}, _log) {
   this.log = _log || createLogger(options);
@@ -622,6 +641,8 @@ function Server (compiler, options = {}, _log) {
     this.listeningApp = http.createServer(app);
   }
 
+  this.options = options;
+
   killable(this.listeningApp);
 
   // Proxy websockets without the initial http request
@@ -803,6 +824,12 @@ Server.prototype.listen = function (port, hostname, fn) {
 
     if (fn) {
       fn.call(this.listeningApp, err);
+    }
+
+    if (this.options.open) {
+      const suffix = (this.options.inline !== false || this.options.lazy === true ? '/' : '/webpack-dev-server/');
+      const uri = createDomain({ host: hostname, port }, this.listeningApp) + suffix;
+      openBrowser(uri, this.options, this.log);
     }
   });
